Cleanup of new EventConsumerHosts
