+++ src/Arcus.EventGrid.Testing/Infrastructure/Hosts/HybridConnectionHost.cs
@@ -1,50 +1,26 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.IO;
 using System.Threading.Tasks;
 using Arcus.EventGrid.Testing.Logging;
 using GuardNet;
 using Microsoft.Azure.Relay;
 using Microsoft.Extensions.Logging;
-using Newtonsoft.Json.Linq;
-using Polly;
 
 namespace Arcus.EventGrid.Testing.Infrastructure.Hosts
 {
-    public class HybridConnectionHost
+    /// <summary>
+    ///     Event consumer host for receiving Azure Event Grid events via Azure Relay with Hybrid Connections
+    /// </summary>
+    public class HybridConnectionHost : EventConsumerHost
     {
-        private static readonly Dictionary<string, string> receivedEvents = new Dictionary<string, string>();
         private readonly HybridConnectionListener _hybridConnectionListener;
-        private readonly ILogger _logger;
 
         private HybridConnectionHost(HybridConnectionListener hybridConnectionListener, ILogger logger)
+            : base(logger)
         {
             Guard.NotNull(hybridConnectionListener, nameof(hybridConnectionListener));
-            Guard.NotNull(logger, nameof(logger));
 
-            _logger = logger;
             _hybridConnectionListener = hybridConnectionListener;
-        }
-
-        /// <summary>
-        ///     Gets the event envelope that includes a requested event (Uses exponential back-off)
-        /// </summary>
-        /// <param name="eventId">Event id for requested event</param>
-        /// <param name="retryCount">Amount of retries while waiting for the event to come in</param>
-        public string GetReceivedEvent(string eventId, int retryCount = 10)
-        {
-            var retryPolicy = Policy.HandleResult<string>(string.IsNullOrWhiteSpace)
-                .WaitAndRetry(retryCount, currentRetryCount => TimeSpan.FromSeconds(Math.Pow(2, currentRetryCount)));
-
-            var matchingEvent = retryPolicy.Execute(() =>
-            {
-                _logger.LogInformation($"Received events are : {string.Join(", ", receivedEvents.Keys)}");
-
-                receivedEvents.TryGetValue(eventId, out var rawEvent);
-                return rawEvent;
-            });
-
-            return matchingEvent;
         }
 
         /// <summary>
@@ -92,6 +68,25 @@ namespace Arcus.EventGrid.Testing.Infrastructure.Hosts
             return new HybridConnectionHost(hybridConnectionListener, logger);
         }
 
+        /// <summary>
+        ///     Stop receiving traffic
+        /// </summary>
+        public async Task Stop()
+        {
+            // Method is added for backwards compatibility reasons
+            await StopAsync();
+        }
+
+        /// <summary>
+        ///     Stop receiving traffic
+        /// </summary>
+        public override async Task StopAsync()
+        {
+            await _hybridConnectionListener.CloseAsync();
+
+            await base.StopAsync();
+        }
+
         private static void HandleReceivedRequest(RelayedHttpListenerContext context, ILogger logger)
         {
             using (var requestStreamReader = new StreamReader(context.Request.InputStream))
@@ -106,25 +101,11 @@ namespace Arcus.EventGrid.Testing.Infrastructure.Hosts
             context.Response.Close();
         }
 
-        /// <summary>
-        ///     Stop receiving traffic
-        /// </summary>
-        public async Task Stop()
-        {
-            _logger.LogInformation("Stopping host");
-            await _hybridConnectionListener.CloseAsync();
-        }
-
         private static void StoreReceivedEvents(string rawEvents, ILogger logger)
         {
             try
             {
-                var parsedEvents = JArray.Parse(rawEvents);
-                foreach (var parsedEvent in parsedEvents)
-                {
-                    var eventId = parsedEvent["Id"]?.ToString();
-                    receivedEvents[eventId] = rawEvents;
-                }
+                EventsReceived(rawEvents);
             }
             catch (Exception)
             {
